+++ linux-gpib/driver/hp_82341/hp_82341_read.c
@@ -44,7 +44,7 @@ int hp_82341_accel_read(gpib_board_t *board, uint8_t *buffer, size_t length, int
 	// read avoids these problems.
 	if(/*tms_priv->holdoff_active == 0 && */length > 1)
 	{
-		int num_bytes;
+		size_t num_bytes;
 		retval = tms9914_read(board, tms_priv, buffer, 1, end, &num_bytes);
 		*bytes_read += num_bytes;
 		if(retval < 0)
@@ -125,7 +125,7 @@ int hp_82341_accel_read(gpib_board_t *board, uint8_t *buffer, size_t length, int
 	// read last byte if we havn't received an END yet
 	if(*end == 0)
 	{
-		int num_bytes;
+		size_t num_bytes;
 		// try to make sure we holdoff after last byte read
 		retval = tms9914_read(board, tms_priv, buffer, length, end, &num_bytes);
 		*bytes_read += num_bytes;
